fixed warnings related to double to float conversion
git-svn-id: https://svn.opendtect.org/od/trunk@25768 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/uiWell/uiwelldahdisplay.cc
@@ -198,8 +198,8 @@ void uiWellDahDisplay::gatherDataInfo( bool first )
     }
     else if ( !zdata_.zistime_ && track() )
     {
-	startpos = track()->getPos( startpos ).z;
-	stoppos = track()->getPos( stoppos ).z;
+	startpos = (float) track()->getPos( startpos ).z;
+	stoppos = (float) track()->getPos( stoppos ).z;
     }
     ld.zrg_.start = startpos;
     ld.zrg_.stop = stoppos;
